Functions/requirements.txt
azure-functions
applicationinsights==0.11.9
azure-common==1.1.23
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
click==7.0
cryptography==2.8
cycler==0.10.0
#datareservoirio==0.7.8
idna==2.8
importlib-resources==1.0.2
joblib==0.14.0
kiwisolver==1.1.0
lxml==4.4.1
matplotlib==3.1.1
numpy==1.14.3 #Was 1.17.3 => numpy updated and is not compatible with alpine. use numpy==1.14.3 or an older version and it should work
oauthlib==3.1.0
pandas==0.25.3
patsy==0.5.1
pyarrow==0.15.1
pycparser==2.19
pyodbc==4.0.27
pyparsing==2.4.4
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
rainflow==2.2.0
requests-oauthlib==1.3.0
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1 #fails to install
six==1.13.0
statsmodels==0.10.1
urllib3==1.25.6
